Preparation for better data structure compatibility with graphics::render().

+++ code/VulkanParticleSystem/graphics.cpp
@@ -921,7 +921,7 @@ namespace graphics {
 		// The command buffer could be in a "pending" state (not finished executing), so we wait for everything to be finished before submission.
 		vkQueueWaitIdle(queue); // TODO: Possible optimisation opportunity here
 
-		vkFreeCommandBuffers(device, commandPool, commandBuffers.size(), commandBuffers.data());
+		vkFreeCommandBuffers(device, commandPool, (uint32_t)commandBuffers.size(), commandBuffers.data());
 		vkDestroyBuffer(device, vertexBuffer, nullptr);
 		vkFreeMemory(device, vertexBufferMemory, nullptr);
 	}
